edited the destroy action in reviews

+++ app/controllers/reviews_controller.rb
@@ -30,8 +30,12 @@ class ReviewsController < ApplicationController
 
     def destroy 
         review = Review.find(params[:id]) 
-        review.destroy
-        render json: review  
+        unless review.nil?
+            review.destroy
+            render json: review
+        else  
+            render json: {error: "Review not found"}, status: 404 
+        end   
     end 
 
     private
